Fix circular dependency and clean up
