+++ examples/webgpu_cubemap_adjustments.html
@@ -31,7 +31,6 @@
 		<script type="module">
 
 			import * as THREE from 'three';
-			import * as Nodes from 'three/nodes';
 
 			import { uniform, mix, cubeTexture, mul, reference, add, positionWorld, normalWorld, modelWorldMatrix, transformDirection, clamp, saturation, hue, reflectVector, context, toneMapping } from 'three/nodes';
