+++ examples/webgpu_cubemap_mix.html
@@ -31,7 +31,6 @@
 		<script type="module">
 
 			import * as THREE from 'three';
-			import * as Nodes from 'three/nodes';
 
 			import { mix, oscSine, timerLocal, cubeTexture, context, float, toneMapping } from 'three/nodes';
